fix(openrouter): decode percentile-object latency/throughput in endpoints API
OpenRouter changed latency_last_30m and throughput_last_30m from bare
numbers to percentile objects {p50,p75,p90,p99}. float64 fields made the
whole-document unmarshal fail, flattenEndpointsResponse fell through all
shape fallbacks, and every provider-panel fetch returned 502
'unrecognized shape'.

New P50Metric decodes number, percentile object, or null and normalizes
to the median (p50) so router priors keep a single scalar.

Verified against live API: 15 endpoints decoded for z-ai/glm-5.3.

+++ internal/openrouter/router.go
@@ -289,7 +289,7 @@ func (r *ProviderRouter) scoreLocked(w RankWeights, ep ProviderEndpoint, stats m
 	// Use API priors when present and local empty.
 	if w.Latency > 0 && ep.LatencyLast30mMs > 0 {
 		if s, ok := stats[ep.ProviderName]; !ok || s.successCount == 0 {
-			ms := ep.LatencyLast30mMs
+			ms := float64(ep.LatencyLast30mMs)
 			if ms < 100 {
 				ms = 100
 			}
@@ -301,7 +301,7 @@ func (r *ProviderRouter) scoreLocked(w RankWeights, ep ProviderEndpoint, stats m
 	}
 	if w.Throughput > 0 && ep.ThroughputLast30mTPS > 0 {
 		if s, ok := stats[ep.ProviderName]; !ok || s.successCount == 0 {
-			t := ep.ThroughputLast30mTPS
+			t := float64(ep.ThroughputLast30mTPS)
 			if t < 1 {
 				t = 1
 			}
